feat：Extract `FailableInsertFeedStoreSpecs` checks into reusable assertions
