+++ EssentialFeedTests/Feed Cache/XCTestCase+FailableInsertFeedStoreSpecs.swift
@@ -1,0 +1,30 @@
+//
+//  XCTestCase+FailableInsertFeedStoreSpecs.swift
+//  EssentialFeedTests
+//
+//  Created by nicaho on 2024/6/14.
+//
+
+import XCTest
+import EssentialFeed
+
+extension FailableInsertFeedStoreSpecs where Self: XCTestCase {
+    func assertThatInsertDeliverErrorsOnInsertionError(on sut: FeedStore, file: StaticString = #filePath, line: UInt = #line) {
+        let feed = uniqueImageFeed().local
+        let timestamp = Date()
+        
+        let insertionError = insert((feed, timestamp), to: sut)
+        
+        XCTAssertNotNil(insertionError, "Expected cache insertion to fail with an error")
+    }
+    
+    func assertThatInsertHasNoSideEffectsOnInsertionError(on sut: FeedStore, file: StaticString = #filePath, line: UInt = #line) {
+        let feed = uniqueImageFeed().local
+        let timestamp = Date()
+        
+        insert((feed, timestamp), to: sut)
+        
+        expect(sut, toRetirve: .empty)
+    }
+    
+}
